mbedtls: cast sockfd to int when storing in net_context fd
The mbedtls server and client bio setup assigned lws_sockfd_type
directly to mbedtls_net_context.fd, which is an int. On Win64
lws_sockfd_type is SOCKET (a 64-bit UINT_PTR), so MSVC warns C4244
"possible loss of data", which fails the build under /WX.

Add an explicit (int) cast, matching the gnutls and openhitls backends
which already cast the same sockfd the same way.

+++ lib/tls/mbedtls/mbedtls-server.c
@@ -351,7 +351,7 @@ lws_tls_server_new_nonblocking(struct lws *wsi, lws_sockfd_type accept_fd)
 		return 1;
 	}
 
-	conn->net.MBEDTLS_PRIVATE_V30_ONLY(fd) = accept_fd;
+	conn->net.MBEDTLS_PRIVATE_V30_ONLY(fd) = (int)accept_fd;
 	mbedtls_ssl_set_bio(&conn->ssl, &conn->net, lws_plat_mbedtls_net_send, lws_plat_mbedtls_net_recv, NULL);
 
 	return 0;
